+++ UI/src/components/features/RipaHomeContainer.vue
@@ -63,15 +63,15 @@ import RipaAddFavoriteDialog from '@/components/molecules/RipaAddFavoriteDialog'
 import RipaApiStopJobMixin from '@/components/mixins/RipaApiStopJobMixin'
 import RipaFavoritesDialog from '@/components/molecules/RipaFavoritesDialog'
 import RipaFormTemplate from '@/components/templates/RipaFormTemplate'
-import RipaHomeContainerMixin from '@/components/mixins/RipaHomeContainerMixin'
 import RipaIntroTemplate from '@/components/templates/RipaIntroTemplate'
 import RipaStatuteDialog from '@/components/molecules/RipaStatuteDialog'
+import RipaStopMixin from '@/components/mixins/RipaStopMixin'
 import { mapGetters, mapActions } from 'vuex'
 
 export default {
   name: 'ripa-home-container',
 
-  mixins: [RipaHomeContainerMixin, RipaApiStopJobMixin],
+  mixins: [RipaStopMixin, RipaApiStopJobMixin],
 
   components: {
     RipaAddFavoriteDialog,
