+++ UI/src/components/mixins/RipaStopMixin.vue
